feat: add centralized logging configuration
