+++ rag_improvement_logging.py
@@ -23,11 +23,6 @@ def setup_improvement_logging():
     if logger.handlers:
         logger.handlers.clear()
     
-    # Clear default root handlers to avoid duplicate logs
-    root_logger = logging.getLogger()
-    if root_logger.handlers:
-        root_logger.handlers.clear()
-    
     # Prevent logs from propagating to the root logger
     logger.propagate = False
