+++ services/simple_conversation_manager.py
@@ -10,7 +10,7 @@ import time
 from typing import Dict, List, Optional
 from services.redis_service import redis_service
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger()
 
 class SimpleConversationManager:
     """
